refactor: css styling adjustment
